+++ CHANGELOG.md
@@ -38,3 +38,4 @@ Fix the bug in the app installer where it fails after unzipping the app file.
 
 - Bugfix installer.py, fixed bug where it did not check for zip to be installed.
 - Added python3 and pip to the list of dependencies to check for.
+- Added a check for venv in the installer.